controller/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
construct==2.10.70
discogs-client==2.3.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-mpd2==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
snapcast==2.3.3
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
